bridge_v2: widen NittyBits to *int32 to accept legacy mongo data
The original elko-Java sometimes wrote nitty_bits with high bits set
(observed on the made: 0x40000008 for one user, 0x40000000 for three
others). NittyBits was declared as *uint8, so reading these records
via BSON unmarshal failed with `1073741832 overflows uint8`,
ensureUserCreated bailed, the ClientSession closed, and the docent
(via pushserver websocketProxy → bridge_v2:1986) disconnected
immediately on connect.

Widen the in-memory and stored type to *int32. The wire/protocol value
is still a single byte; future writers should mask & 0xFF when
serializing back to clients.

  - schema.go:    NittyBits *uint8  →  *int32
  - utils.go:     add Int32P helper to mirror Uint8P/Uint32P
  - client_session.go: ensureUserCreated uses Int32P(0) for new users

Also fixed in mongo on the made (4 affected user records masked to
their low byte) so existing data loads without the type widening too.

+++ bridge_v2/bridge/schema.go
@@ -208,7 +208,11 @@ type HabitatMod struct {
 	MagicType       *uint8    `json:"magic_type,omitempty" bson:"magic_type,omitempty"`
 	Mass            *uint8    `json:"mass,omitempty" bson:"mass,omitempty"`
 	Neighbors       *[]string `json:"neighbors,omitempty" bson:"neighbors,omitempty"`
-	NittyBits       *uint8    `json:"nitty_bits,omitempty" bson:"nitty_bits,omitempty"`
+	// NittyBits is *int32 rather than *uint8 because legacy elko-Java
+	// data sometimes stored values >255 (sentinels like 0x40000008 mixed
+	// in with the nominal flag byte). Wire/protocol use is still single-byte
+	// — callers should mask & 0xFF when serializing back to clients.
+	NittyBits       *int32    `json:"nitty_bits,omitempty" bson:"nitty_bits,omitempty"`
 	// Noid is uint16 (not uint8) because Elko sends 256 as a sentinel for
 	// "the session user's own objects" — Avatar, Head, Paper, Tokens —
 	// before the bridge has assigned a local noid. See UNASSIGNED_NOID.
